Potential crash when a connection is stopped
Integrate Wakaama fix for issue #257
When a connection to the server results in a connection failure to the
Device Management server, a connection is initiated to the bootstrap
server. Then a connection is initiated to the Device Management server.
When the connection is stopped, a potential crash could happen because
list of connection was not correctly released.

Resolves: LE-6580
Change-Id: I6d6fdb71db7e57e126d24fcf89ca3b3e3ccc5ff6