Optimize connection to EDM server
The EDM server lifetime value will not expire. To minimize the amount of data transfered
between the client and the server, we will not DEREGISTER the serverID for the EDM server.
This will ensure that subsequent connections to the EDM server will only perform a REGUPDATE
instead of a REGISTER.

Resolves: LE-16034
Change-Id: I1e96d55eb91914eccdea5d2040f44cde99d130c1