Improvement on block1 data receipt and retransmission
When data are received from the AV server using block1, the device can
answer with some delay in order to forward the data from wakaama to
application layers (AVC, ATSWI).
Then a 1st retransmission can be received while the initial data block
was already treated.
For this retransmission, the initial response should be automatically
sent again by wakaama and not forwarded to the application layers
(with the application layers response).

Resolves: LE-16928
Change-Id: Icea14624211ba8ff1f170d613ec68079abe3a486
Relates: ATSWI-558